refactor(StationEnvironmentElement): refactor file structure
