+++ src/main/scala/ulisse/applications/useCases/SimulationService.scala
@@ -6,7 +6,7 @@ import ulisse.applications.ports.SimulationPorts
 import ulisse.entities.route.RouteEnvironmentElement
 import ulisse.entities.simulation.data.{Engine, SimulationData}
 import ulisse.entities.simulation.environments.railwayEnvironment.{ConfigurationData, RailwayEnvironment}
-import ulisse.entities.station.StationEnvironments.StationEnvironmentElement
+import ulisse.entities.station.StationEnvironmentElement
 import ulisse.entities.timetable.DynamicTimetables.DynamicTimetable
 import ulisse.entities.train.TrainAgents.TrainAgent
 import ulisse.infrastructures.commons.TimeProviders.*
