+++ src/main/scala/ulisse/entities/route/Tracks.scala
@@ -24,10 +24,6 @@ object Tracks:
         case TrackDirection.Forward  => TrackDirection.Backward
         case TrackDirection.Backward => TrackDirection.Forward
 
-  /** Errors for Track */
-  enum Errors extends BaseError:
-    case InvalidTrackId
-
   /** Defines a rail in a Route. */
   trait Track extends TrainAgentsContainer[Track]:
     /** Try to put train inside the track given the desired direction */
@@ -47,16 +43,21 @@ object Tracks:
 
   /** Factory for [[Track]] instances. */
   object Track:
-
+    /** Minimum track id. */
+    val minPlatformId: Int = 1
     /** Creates a `Track` instance. If track id is not positive it is set to 1 */
     def apply(id: Int)(using minPermittedDistanceBetweenTrains: Double): Track =
-      TrackImpl(math.max(1, id), Seq(), None)
+      TrackImpl(math.max(minPlatformId, id), Seq(), None)
 
     /** Creates a `Track` instance with validation. If track id is not positive an error is returned */
     def createCheckedTrack(
         id: Int
     )(using minPermittedDistanceBetweenTrains: Double): Either[NonEmptyChain[Tracks.Errors], Track] =
       validatePositive(id, Errors.InvalidTrackId).toValidatedNec.toEither.map(Track(_))
+
+    /** Represents errors that can occur during `Tracks` creation. */
+    enum Error extends BaseError:
+      case InvalidTrackId
 
     private final case class TrackImpl(id: Int, trains: Seq[TrainAgent], currentDirection: Option[TrackDirection])(
         using val minPermittedDistanceBetweenTrains: Double
